new changes

+++ scrape.py
@@ -1,7 +1,8 @@
 import pandas as pd
 import os
 
-os.remove('data.json')
+if (os.path.exists('data.json')):
+    os.remove('data.json')
 dfs = pd.read_html('https://www.nepalipaisa.com/StockLive.aspx')
 len(dfs)
 df = dfs[0]
@@ -9,4 +10,5 @@ df.head
 df
 df.drop(['LTV','Quantity','Difference Rs.','No of Transaction'], axis=1, inplace=True)
 df.rename(columns={'Closing Price' : 'Closing', 'Max Price' : 'High', 'Min Price' : 'Low', 'Opening Price' : 'Open', '%Change':'Change'}, inplace=True)
+print(df)
 df.to_excel('data.xlsx')
